refactor src.vex.bumper_switch_sensor & .distance_sensor
