+++ src/vex/distance_sensor/__init__.py
@@ -5,12 +5,12 @@ from collections.abc import Sequence
 
 from abm.decor import act, sense
 
-from ._abstract_device import Device
-from .brain.port import Ports
-from .units_common import DistanceUnits
+from .._abstract_device import Device
+from ..brain.port import Ports
+from ..units_common import DistanceUnits
 
 # pylint: disable=unused-import
-from .util.doc import robotmesh_doc, vexcode_doc   # noqa: F401
+from ..util.doc import robotmesh_doc, vexcode_doc   # noqa: F401
 
 
 __all__: Sequence[str] = ('Sonar',)
